Prepare coefficients part of the latex table

Prepare coefficients part of the latex table

light_table_coefficients(
  object,
  ncols_models,
  type,
  coeff_data,
  order_variable,
  omit,
  covariate.labels,
  reference_level_position,
  rules_between_covariates
)

Arguments

object
List of object

ncols_models
Number of columns

type
Reporting output used. Accepted values are latex and html

coeff_data
Output from secoeff

order_variable
A vector that indicates the order in which variables will appear in the output.

omit
List of variables that should be removed from the table

covariate.labels
A character vector of labels for columns in regression tables. Their layout, in terms of the number of columns associated with each label, is given by the argument column.separate.

reference_level_position
In multinomial logit, coefficients are relative to the reference level. A column can be added to represent an empty coefficient modality that is indicated as being the reference. If NULL, this feature is ignored. Otherwise, the column is placed at the position indicated by the argument

rules_between_covariates
A numeric vector that specifies how rules should be laid out across rows of the regression table